:fire: Remove SA related content
