+++ lamindb/dev/__init__.py
@@ -3,22 +3,10 @@
 .. autosummary::
    :toctree: .
 
-   db
    LazyDataFrame
    datasets
-
-Utilities:
-
-.. autosummary::
-   :toctree: .
-
-   doc_args
 """
-
-from lnschema_core.dev import id  # noqa
 
 from lamindb.dev.storage.object._lazy_field import Lazy as LazyDataFrame
 
 from . import datasets  # noqa
-from . import db  # noqa
-from ._docs import doc_args  # noqa
