--- lamindb/dev/db/__init__.py
@@ -1,14 +1,0 @@
-"""Helpers for the db API the database.
-
-Select:
-
-.. autosummary::
-   :toctree: .
-
-   SelectStmt
-   ExecStmt
-"""
-
-from ._add import add
-from ._select import ExecStmt, SelectStmt, select
-from ._session import Session
