--- lamindb/schema/dev/__init__.py
@@ -1,6 +1,0 @@
-from lnschema_core import dev as _dev
-from lnschema_core.dev import *  # noqa
-
-__doc__ = _dev.__doc__
-
-from . import id  # noqa
