--- lamindb/schema/dev/id.py
@@ -1,4 +1,0 @@
-from lnschema_core.dev import id as _id
-from lnschema_core.dev.id import *  # noqa
-
-__doc__ = _id.__doc__
