--- lamindb/schema/dev/sqlmodel.py
@@ -1,4 +1,0 @@
-from lnschema_core.dev import sqlmodel as _sqlmodel
-from lnschema_core.dev.sqlmodel import *  # noqa
-
-__doc__ = _sqlmodel.__doc__
