+++ tests/conftest.py
@@ -3,7 +3,8 @@ import pytest
 
 
 def pytest_sessionstart(session: pytest.Session):
-    if lamindb_setup._USE_DJANGO:
+    _USE_DJANGO = True
+    if _USE_DJANGO:
         lamindb_setup.init(storage="./lamindb-unit-tests")
     else:
         lamindb_setup.init(storage="./lamindb-unit-tests", schema="bionty")
